+++ lib/inspec/profile.rb
@@ -271,6 +271,15 @@ module Inspec
       end
       res[:sha256] = sha256
       res[:parent_profile] = parent_profile unless parent_profile.nil?
+
+      # convert legacy os-* supports to their platform counterpart
+      if res[:supports] && !res[:supports].empty?
+        res[:supports].each do |support|
+          support[:"platform-family"] = support.delete(:"os-family") if support.key?(:"os-family")
+          support[:"platform-name"] = support.delete(:"os-name") if support.key?(:"os-name")
+        end
+      end
+
       res
     end